LibMarkdown: Ability to configure Render Extensions
This patch implements an adhoc infrastructure for configuring Render
Extensions for each call to the `render_to_html()` functions.

The fragment links in heading feature is now under an Render Extension
`FragmentLinksInHeading` and the prepending of `file://` to absolute
paths is now under `PrependFileProtocolIfAbsolutePath`. Both of these
extensions are enabled by default.

The CommonMark test suite does not expect any of these extensions, so,
TestCommonmark now disables all extensions. This makes sure that some
tests don't fail just because of an extension (they otherwise succeed if
all the extensions are disabled)